requirements-pi.txt
# Minimal dependencies for Raspberry Pi trading bot
# Tested on Raspberry Pi 4 with Python 3.9+

# Core trading library
ccxt==4.1.10

# Async support
aiohttp==3.9.1
websockets==12.0

# Numerical computing (lightweight alternatives)
numpy==1.24.3
pandas==2.0.3  # For tax calculations

# System monitoring
psutil==5.9.6
prometheus-client==0.19.0

# HTTP requests for webhooks
requests==2.31.0

# Configuration
python-dotenv==1.0.0  # Para cargar .env files
pyyaml==6.0.1

# Tax reporting (optional but recommended)
rich==13.7.0  # For tax dashboard - can be omitted if not using dashboard

# Optional: For better performance on Pi
# numexpr==2.8.7  # Faster numpy operations
# bottleneck==1.3.7  # Optimized array functions